requirements.txt
# 
# !!! Requirements file generated by nmk-base plugin, don't edit !!!
#

#
# Package dependencies from nmk project files
#

nmk-badges
nmk-base
nmk-doc
nmk-github
nmk-python
nmk-vscode
sphinx-autoapi


